refactor(scripts): lift the adwaita element reader out
Three scripts answered "which custom elements does adwaita-web ship" and
three answered it differently — this one by scanning customElements.define
recursively, the other two by listing adw-*.ts filenames in one directory.
Same CI job, 65 vs 50, and the smaller number fed the published matrix.

The reader moves to scripts/adwaita-elements.mjs as a flat sibling, the way
posix-path-slice-exceptions.mjs is imported by its check: every other module
in scripts/ is flat, and the one directory there (manifest-conformance/) is a
multi-file registry with its own README, not a helpers bin.

No behaviour change yet: this check still reports the same 65 tags.
